+++ R/Gargle-class.R
@@ -22,7 +22,7 @@
 #' Defaults to the option named `"gargle_oauth_email"`, retrieved by
 #' [gargle::gargle_oauth_email()] (unless a wrapper package implements different
 #' default behavior).
-#' @param app A Google OAuth client, preferably constructed via
+#' @param client A Google OAuth client, preferably constructed via
 #'   [gargle::gargle_oauth_client_from_json()], which returns an instance of
 #'   `gargle_oauth_client`. For backwards compatibility, for a limited time,
 #'   gargle will still accept an "OAuth app" created with [httr::oauth_app()].
@@ -39,6 +39,8 @@
 #' @inheritParams httr::oauth2.0_token
 #' @param ... Absorbs arguments intended for use by other credential functions.
 #'   Not used.
+#' @param app `r lifecycle::badge('deprecated')` Replaced by the `client`
+#'   argument.
 #' @return An object of class [Gargle2.0], either new or loaded from the cache.
 #' @export
 #' @examples
@@ -46,7 +48,7 @@
 #' gargle2.0_token()
 #' }
 gargle2.0_token <- function(email = gargle_oauth_email(),
-                            app = gargle_client(),
+                            client = gargle_client(),
                             package = "gargle",
                             ## params start
                             scope = NULL,
@@ -54,7 +56,17 @@ gargle2.0_token <- function(email = gargle_oauth_email(),
                             ## params end
                             credentials = NULL,
                             cache = if (is.null(credentials)) gargle_oauth_cache() else FALSE,
-                            ...) {
+                            ...,
+                            app = deprecated()) {
+  if (lifecycle::is_present(app)) {
+    lifecycle::deprecate_soft(
+      "1.5.0",
+      "gargle2.0_token(app)",
+      "gargle2.0_token(client)"
+    )
+    client <- app
+  }
+
   params <- list(
     scope = scope,
     use_oob = use_oob,
@@ -62,9 +74,9 @@ gargle2.0_token <- function(email = gargle_oauth_email(),
   )
 
   # pseudo-oob flow
-  client_type <- if (inherits(app, "gargle_oauth_client")) app$type else NA
+  client_type <- if (inherits(app, "gargle_oauth_client")) client$type else NA
   if (use_oob && identical(client_type, "web")) {
-    params$oob_value <- select_pseudo_oob_value(app$redirect_uris)
+    params$oob_value <- select_pseudo_oob_value(client$redirect_uris)
   }
   # params$oob_value is deliberately left unspecified for conventional oob,
   # with the intent of falling back to urn:ietf:wg:oauth:2.0:oob
@@ -80,7 +92,7 @@ gargle2.0_token <- function(email = gargle_oauth_email(),
 
   Gargle2.0$new(
     email = email,
-    app = app,
+    client = client,
     package = package,
     params = params,
     credentials = credentials,
@@ -108,8 +120,21 @@ gargle2.0_token <- function(email = gargle_oauth_email(),
 #' directory of such files. In contrast, `Token2.0` tokens are cached as
 #' components of a list, which is typically serialized to `./.httr-oauth`.
 #'
-#' @param email Optional email address. See [gargle2.0_token()] for full details.
+
+#' @param email Optional email address. See [gargle2.0_token()] for full
+#'   details.
+#' @param client An OAuth consumer application.
 #' @param package Name of the package requesting a token. Used in messages.
+#' @param credentials Exists largely for testing purposes.
+#' @param params A list of parameters for the internal function
+#'   `init_oauth2.0()`, which is a modified version of [httr::init_oauth2.0()].
+#'   gargle actively uses `scope` and `use_oob`, but does not use `user_params`,
+#'   `type`, `as_header` (hard-wired to `TRUE`), `use_basic_auth` (accept
+#'   default of `use_basic_auth = FALSE`), `config_init`, or
+#'   `client_credentials`.
+#' @param cache_path Specifies the OAuth token cache. Read more in
+#'   [gargle::gargle_oauth_cache()].
+#' @param app `r lifecycle::badge('deprecated')` Use `client` instead.
 #'
 #' @keywords internal
 #' @export
@@ -119,29 +144,33 @@ Gargle2.0 <- R6::R6Class("Gargle2.0", inherit = httr::Token2.0, list(
   email = NULL,
   #' @field package Name of the package requesting a token. Used in messages.
   package = NULL,
+  #' @field client An OAuth client.
+  client = NULL,
   #' @description Create a Gargle2.0 token
-  #' @param app An OAuth consumer application.
-  #' @param credentials Exists largely for testing purposes.
-  #' @param params A list of parameters for the internal function
-  #'   `init_oauth2.0()`, which is a modified version of
-  #'   [httr::init_oauth2.0()]. gargle actively uses `scope` and `use_oob`, but
-  #'   does not use `user_params`, `type`, `as_header` (hard-wired to `TRUE`),
-  #'   `use_basic_auth` (accept default of `use_basic_auth = FALSE`),
-  #'   `config_init`, or `client_credentials`.
-  #' @param cache_path Specifies the OAuth token cache. Read more in
-  #'   [gargle::gargle_oauth_cache()].
   #' @return A Gargle2.0 token.
   initialize = function(email = gargle_oauth_email(),
-                        app = gargle_client(),
+                        client = gargle_client(),
                         package = "gargle",
                         credentials = NULL,
                         params = list(),
-                        cache_path = gargle_oauth_cache()) {
+                        cache_path = gargle_oauth_cache(),
+                        app = deprecated()) {
     gargle_debug("Gargle2.0 initialize")
+    # I'm using deprecate_warn() intentionally here. Most folks should be
+    # instantiating through gargle2.0_token() anyway, so anyone who sees this
+    # warning probably needs to see it.
+    if (lifecycle::is_present(app)) {
+      lifecycle::deprecate_warn(
+        "1.5.0",
+        "Gargle2.0$initialize(app)",
+        "Gargle2.0$initialize(client)"
+      )
+      client <- app
+    }
     stopifnot(
       is.null(email) || is_scalar_character(email) ||
         isTRUE(email) || isFALSE(email) || is_na(email),
-      is.oauth_app(app),
+      is.oauth_app(client),
       is_string(package),
       is.list(params)
     )
@@ -163,7 +192,10 @@ Gargle2.0 <- R6::R6Class("Gargle2.0", inherit = httr::Token2.0, list(
 
     self$endpoint   <- gargle_oauth_endpoint()
     self$email      <- email
-    self$app        <- app
+    self$client     <- client
+    # for backwards compatibility and also because the parent class has $app;
+    # I can never remove it
+    self$app        <- client
     self$package    <- package
     params$scope    <- normalize_scopes(add_email_scope(params$scope))
     params$query_authorize_extra <- list(login_hint = login_hint)
@@ -192,7 +224,7 @@ Gargle2.0 <- R6::R6Class("Gargle2.0", inherit = httr::Token2.0, list(
   format = function(...) {
     x <- list(
       oauth_endpoint = "google",
-      app            = self$app$appname,
+      client         = self$client$name,
       email          = cli::format_inline("{.email {self$email}}"),
       scopes         = commapse(base_scope(self$params$scope)),
       credentials    = commapse(names(self$credentials))
@@ -303,10 +335,10 @@ encourage_httpuv <- function() {
   invisible()
 }
 
-# I want to encourage users to create an OAuth app (older httr-y language) or
-# client (newer httr2-y language) directly from downloaded JSON, using
-# gargle_oauth_client_from_json(). Sometimes there are multiple URIs and I think
-# we can usually figure out which one to use for the pseudo-oob flow.
+# I want to encourage users to create an OAuth client (newer httr2-y language)
+# directly from downloaded JSON, using gargle_oauth_client_from_json().
+# Sometimes there are multiple URIs and I think we can usually figure out which
+# one to use for the pseudo-oob flow.
 select_pseudo_oob_value <- function(redirect_uris) {
   # https://developers.google.com/identity/protocols/oauth2/resources/oob-migration#inspect-your-application-code
   bad_values <- c(
@@ -331,8 +363,8 @@ select_pseudo_oob_value <- function(redirect_uris) {
 
   if (length(redirect_uris) == 0) {
     gargle_abort('
-      OAuth client (a.k.a "app") does not have a redirect URI suitable for \\
-      the pseudo-OOB flow.')
+      OAuth client does not have a redirect URI suitable for the pseudo-OOB \\
+      flow.')
   }
 
   if (length(redirect_uris) > 1) {